Move migration test to `sozo`
