+++ crates/dojo-world/Cargo.toml
@@ -27,5 +27,4 @@ url = "2.2.2"
 [dev-dependencies]
 assert_fs = "1.0.9"
 dojo-test-utils = { path = "../dojo-test-utils" }
-dojo-lang = { path = "../dojo-lang" }
 tokio = { version = "1.28.0", features = [ "full" ] }
